Rename to citadel: minimum move and imports change
This change only moves files and updates import paths.  It does
change the behavior of any APIs.  The code continues to compile.
