+++ cmd/citadel/citadel.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"log"
+
+	"github.com/enj/citadel/pkg/cmd/citadel"
+)
+
+func main() {
+	if err := citadel.Execute(); err != nil {
+		log.Fatal(err)
+	}
+}
